Module photon::channels

Functions

inc_blue_channel
Increment every pixel's Blue channel by a constant.

inc_channel
Alter a select channel by incrementing its value by a constant.

inc_green_channel
Increment every pixel's Green channel by a constant.

inc_red_channel
Increment every pixel's Red channel by a constant.

inc_two_channels
Increment two channels' values simultaneously by adding an offset to each channel per pixel.

remove_blue_channel
Remove the Blue channel's influence in an image, by setting its value to zero.

remove_channel
Set a certain channel to zero, thus removing the channel's influence in the pixels' final rendered colour.

remove_green_channel
Remove the Green channel's influence in an image, by setting its value to zero.

remove_red_channel
Remove the Red channel's influence in an image, by setting its value to zero.