Module photon::colour_spaces

Image manipulation effects in HSL, LCh and HSV.

Functions

darken_hsl
Darken the image by a specified amount in the HSL colour space.

darken_hsv
Darken the image's colours by a specified amount in the HSV colour space.

darken_lch
Darken the image by a specified amount in the LCh colour space.

desaturate_hsl
Desaturate the image by a specified amount in the HSL colour space.

desaturate_hsv
Desaturate the image by a specified amount in the HSV colour space.

desaturate_lch
Desaturate the image by a specified amount in the LCh colour space.

hsl
Image manipulation effects in the HSL colour space.

hsv
Image manipulation in the HSV colour space.

hue_rotate_hsl
Shift hue by a specified number of degrees in the HSL colour space.

hue_rotate_hsv
Shift hue by a specified number of degrees in the HSV colour space.

hue_rotate_lch
Shift hue by a specified number of degrees in the LCh colour space.

lch
Image manipulation effects in the LCh colour space

lighten_hsl
Lighten an image by a specified amount in the HSL colour space.

lighten_hsv
Lighten an image by a specified amount in the HSV colour space.

lighten_lch
Lighten an image by a specified amount in the LCh colour space.

saturate_hsl
Increase the image's saturation by converting each pixel's colour to the HSL colour space and increasing the colour's saturation.

saturate_hsv
Increase the image's saturation in the HSV colour space.

saturate_lch
Increase the image's saturation in the LCh colour space.