Function photon::colour_spaces::hue_rotate_hsl

pub fn hue_rotate_hsl(img: &mut PhotonImage, degrees: f32)

Shift hue by a specified number of degrees in the HSL colour space.

Arguments

img - A PhotonImage.

mode - The number of degrees to shift the hue by, or hue rotate by.

Example

// For example to hue rotate/shift the hue by 120 degrees in the HSL colour space:
use photon::color_spaces::hue_rotate_hsl;
 
// Open the image. A PhotonImage is returned.
let img: PhotonImage = open_image("images/flowers.PNG");
 
hue_rotate_hsl(&mut img, 120);