Module photon::conv

Convolution effects such as sharpening, blurs, sobel filters, etc.,

Functions

box_blur
Apply a box blur effect.

detect_45_deg_lines
Detect lines at a forty five degree angle in an image, and highlight these only.

detect_135_deg_lines
Detect lines at a 135 degree angle in an image, and highlight these only.

detect_horizontal_lines
Detect horizontal lines in an image, and highlight these only.

detect_vertical_lines
Detect vertical lines in an image, and highlight these only.

edge_detection
Apply edge detection to an image, to create a dark version with its edges highlighted.

edge_one
Preset edge effect.

emboss
Apply an emboss effect to an image.

gaussian_blur
Apply a gaussian blur effect.

identity
Apply an identity kernel convolution to an image.

laplace
Apply a standard laplace convolution.

noise_reduction
Noise reduction.

prewitt_horizontal
Apply a horizontal Prewitt convolution to an image.

sharpen
Sharpen an image.

sobel_horizontal
Apply a horizontal Sobel filter to an image.

sobel_vertical
Apply a vertical Sobel filter to an image.