Module photon::effects

Special effects.

Functions

colorize
Colorizes the green channels of the image.

create_gradient_map
Create a gradient map between two RGB colours.

duotone
Pass the image through a duotone filter (comprising two colours, and their gradient from one colour to the other). This is similar to greyscaling an image, but rather than having the gradient transition from black to white, it should be between two other colours, red to green, for example.

halftone
Halftoning effect.

inc_brightness
Increase the brightness of an image by a factor.

multiple_offsets
Adds multiple offsets to the image by a certain number of pixels (on two channels).

offset
Adds an offset to the image by a certain number of pixels.

offset_blue
Adds an offset to the blue channel by a certain number of pixels.

offset_green
Adds an offset to the green channel by a certain number of pixels.

offset_red
Adds an offset to the red channel by a certain number of pixels.

primary
Reduces an image to the primary colours.

solarize
Applies a solarizing effect to an image.

tint
Tint an image by adding an offset to averaged RGB channel values.