Module photon::filters

Functions

bluechrome
Blue monochrome effect.

diamante
Custom filter with a blue/turquoise tint.

flagblue
Royal blue tint

islands
Custom filter with an aquamarine tint.

liquid
Blue-inspired tint.

lix
Solarization on the Red and Green channels.

marine
Add a green/blue mixed hue to an image.

mauve
Purple-infused filter.

neue
Solarization on the Blue channel.

oceanic
Add an aquamarine-tinted hue to an image.

perfume
Increase the blue channel, with moderate increases in the Red and Green channels.

radio
Fallout-style radio effect.

rosetint
Rose-tinted filter.

ryo
Solarization on the Red and Blue channels.

seagreen
Dark green hue, with tones of blue.

serenity
Custom filter with an increase in the Blue channel's values.

twenties
Slight-blue tinted historical effect.

vintage
Vintage filter with a red tint.