Module photon::multiple

Image manipulation with multiple images, including adding watermarks, changing backgrounds, etc.,

Functions

blend
Blend two images together. The blend_mode (3rd param) determines which blending mode to use; change this for varying effects. The blend modes available include: overlay, over, atop, xor, multiply, burn, soft_light, hard_light difference, lighten, darken, dodge, plus, exclusion (more to come) NOTE: The first image must be smaller than the second image passed as params. If the first image were larger than the second, then there would be overflowing pixels which would have no corresponding pixels in the second image.

replace_background
Change the background of an image (using a green screen/color screen).

watermark
Add a watermark to an image.