Module photon::channels

Channel manipulation.

Functions

alter_blue_channel
Increment or decrement every pixel's Blue channel by a constant.

alter_channel
Alter a select channel by incrementing or decrementing its value by a constant.

alter_channels
Increment all 3 channels' values by adding an amt to each channel per pixel.

alter_green_channel
Increment or decrement every pixel's Green channel by a constant.

alter_red_channel
Increment or decrement every pixel's Red channel by a constant.

alter_two_channels
Increment/decrement two channels' values simultaneously by adding an amt to each channel per pixel.

color_sim
Get the similarity of two colours in the lab colour space using the CIE76 formula.

remove_blue_channel
Remove the Blue channel's influence in an image.

remove_channel
Set a certain channel to zero, thus removing the channel's influence in the pixels' final rendered colour.

remove_green_channel
Remove the Green channel's influence in an image.

remove_red_channel
Remove the Red channel's influence in an image.

selective_desaturate
Selectively desaturate pixel colours which are similar to the reference colour provided.

selective_greyscale
Selectively changes a pixel to greyscale if it is not visually similar or close to the colour specified. Only changes the colour of a pixel if its RGB values are within a specified range.

selective_hue_rotate
Selective hue rotation.

selective_lighten
Selectively lighten an image.

selective_saturate
Selectively saturate pixel colours which are similar to the reference colour provided.

swap_channels
Swap two channels.